feat: 시청한 영상에서 시청률 정수로 반환

+++ src/main/java/app/allstackproject/privideo/dto/video/HistoryResponse.java
@@ -42,7 +42,7 @@ public class HistoryResponse {
         private Long id;
         private String name;
         private String img;
-        private double watchRate;
+        private Long watchRate;
         private LocalDateTime recentWatch;
     }
 }
